Validate the Qwen3 embedder against real weights, and fix what that found
Ran tests/test_embedding_live.py against mlx-community/Qwen3-Embedding-0.6B-8bit on
Apple Silicon. Two things came out of it.

Vectors were not unit length. The model runs in bfloat16, where a sum of squares over
1024 dimensions carries about three decimal digits, so normalising in the model dtype
left them off by ~2.5e-4 - and InMemoryPolicyIndex scores with a plain dot product on
the assumption that they are unit vectors. Normalisation now happens in float32.

The claim I had written about the placeholder was wrong. The live test asserted that
HashingEmbedder ranks a lexically similar distractor above the clause that answers the
query, and on the fixture I had chosen it does not: shared tokens ("stage", "risk")
carry it to the right answer, 0.257 against 0.063. Replaced with a query that shares no
content word with the clause it asks about, where the placeholder does rank the wrong
clause first (0.058 against 0.142) and Qwen3-Embedding ranks correctly (0.736 against
0.698). The limitation is now pinned by a test in the offline suite rather than
asserted in a comment.

Worth recording: neither embedder gets every paraphrase right. A "worsening outlook /
lifetime loss provisioning" query ranks the collateral clause first under both. The
0.6B embedder is a floor, not a solution, and retrieval quality still needs measuring
against a real corpus.

Also: a cohort plan reaching /v1/factsheet built CreditFactsheet with obligor_id=None
and surfaced a raw pydantic ValidationError. The factsheet is obligor-level throughout
and every consumer of it reasons about one borrower, so build_factsheet now refuses a
portfolio plan and points at /v1/query/fetch.

And COUNT(DISTINCT obligor_id) was emitted without the identifier clearing the column
allowlist. It is in the registry, so the SQL was valid, but the check was being skipped.

185 tests offline, 7 live embedder tests passing.

+++ credit-risk-finetuning/src/credit_risk/rag/embedding.py
@@ -140,7 +140,11 @@ class MLXEmbedder:
             # condition. Dropping it would embed the heading and discard the rule.
             ids = ids[-self.max_tokens:]
         hidden = self._model.model(mx.array([ids]))
-        pooled = hidden[0, -1, :]
+        # Normalise in float32. The model runs in bfloat16, where sum-of-squares over a
+        # thousand dimensions carries about three decimal digits, so normalising in the
+        # model dtype left vectors off unit length by ~2.5e-4. InMemoryPolicyIndex scores
+        # with a plain dot product on the assumption that they are unit vectors.
+        pooled = hidden[0, -1, :].astype(mx.float32)
         norm = mx.sqrt(mx.sum(pooled * pooled))
         if float(norm) > 0:
             pooled = pooled / norm
